Django-Middleware-0x03/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-nested-routers==0.94.2
mysqlclient==2.2.7
PyJWT==2.9.0
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.13.2
